+++ src/Deprecated.jl
@@ -5,6 +5,8 @@
 ## Remove below before v0.17
 ## ======================================================================================================
 
+@deprecate getPointRepr(x::HomotopyDensity) getPointType(x)
+
 @deprecate getInfoPerCoord(mkd::HomotopyDensity, aspartial::Bool = true) getObservability(mkd, aspartial)
 
 @deprecate exists_BTLabel(hode::HomotopyDensity, idx::Int) isassigned(hode, idx)