+++ NEWS.md
@@ -12,6 +12,7 @@ Major news relating to breaking changes in ApproxManifoldProducts.jl
 - Rename `HomotopyDensity.structure` instead of `.geometric_permute` (breaking change)
 - Drop `HomotopyDensity.leaf_kernels`, switch to new approach with `.minors_detail` (breaking change)
 - Adding `HomotopyDensity.majors_*` fields 
+- Deprecation, use `getPointType` instead of `getPointRepr`. (breaking change)
 
 ## v0.14 (26Q2)